backend/requirements.txt
# ===== VidyaMind Python Backend — Dependencies =====
#
# Install with:  pip install -r requirements.txt
#
# Core:
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.1.0

# AI / LLM:
groq==0.15.0

# Utilities (already included with FastAPI, listed for clarity):
pydantic>=2.0
